+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.6 Developing
 * Bug fixed in DateInput
+* ReCaptcha component added.
 
 ## 1.1.5 
 * was update RadioInput css.